remove the mozilla specific processor class
merge the mozilla specific processor algorithm class into the base class, and combine all the transform rules into one set in the process
